Move HeroInfo and TownInfo structures back to cpp file
